% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commits.R
\name{plot_commits}
\alias{plot_commits}
\title{Plot GitHub commits as clock or bar chart}
\usage{
plot_commits(df, crit = "author", type = "clock")
}
\arguments{
\item{df}{Data frame from get_commits(), must contain datetime column \code{date} or \code{dttm}}

\item{crit}{Character. Column name to group/colour by (default "author")}

\item{type}{Character. "clock" or "bar" (default "clock")}
}
\value{
ggplot object
}
\description{
Plot GitHub commits as clock or bar chart
}
\examples{
\dontrun{
library(getcommit)
library(clockplot)

# Fetch the most recent 50 commits from tidyverse/dplyr
comm <- get_commits("tidyverse/dplyr", n = 50)

# Plot as clockplot grouped by author
plot_commits(comm, crit = "author", type = "clock")

# Plot as bar chart grouped by author
plot_commits(comm, type = "bar")
}
}
