% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_dt_tm.R
\name{sep_dt_tm}
\alias{sep_dt_tm}
\title{Separate datetime into date and time columns}
\usage{
sep_dt_tm(df, datetime_col = "date")
}
\arguments{
\item{df}{A data frame containing a datetime column (POSIXct) named 'dttm'}

\item{datetime_col}{Name of the datetime column, default 'dttm'}
}
\value{
Data frame with separate 'dt' (Date) and 'tm' (hms) columns
}
\description{
Separate datetime into date and time columns
}
\examples{
comm <- get_commits("tidyverse/dplyr", n = 10)
comm2 <- sep_dt_tm(comm)
}
