% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_commits.R
\name{summarize_commits}
\alias{summarize_commits}
\title{Summarize commit activity}
\usage{
summarize_commits(df, by = c("author", "year", "month", "day", "hour"))
}
\arguments{
\item{df}{A tibble returned by \code{\link[=get_commits]{get_commits()}}.}

\item{by}{A character vector specifying grouping variables.
Choices are \code{"author"}, \code{"year"}, \code{"month"}, \code{"day"}, or \code{"hour"}.}
}
\value{
A tibble with counts per group.
}
\description{
Aggregates commit counts by one or more grouping variables.
}
\examples{
\dontrun{
comm <- get_commits("tidyverse/dplyr", n = 50)
summarize_commits(comm, by = "author")
summarize_commits(comm, by = c("author", "year"))
}

}
